requirements.txt
flask
mercadopago
qrcode
pillow
python-escpos
python-dotenv